requirements.txt
python-docx
requests
google-genai
streamlit